feat(pragma_consumer): Moved stuff - not sure for the types yet
