+++ pragma-common/src/errors.rs
@@ -1,0 +1,8 @@
+#[derive(Debug)]
+pub enum ConversionError {
+    FailedSerialization,
+    InvalidDateTime,
+    BigDecimalConversion,
+    FeltConversion,
+    U128Conversion,
+}
