+++ pragma-common/src/lib.rs
@@ -1,2 +1,3 @@
+pub mod errors;
 pub mod tracing;
 pub mod types;
