+++ pragma-common/src/types/merkle_tree.rs
@@ -1,9 +1,13 @@
 use std::cmp::Ordering;
 
-use color_eyre::eyre::eyre;
-use color_eyre::eyre::Result;
-use starknet_crypto::pedersen_hash;
-use starknet_crypto::Felt;
+use starknet::core::{crypto::pedersen_hash, types::FieldElement};
+use thiserror::Error;
+
+#[derive(Debug, Error)]
+pub enum MerkleTreeError {
+    #[error("cannot build a merkle tree from empty leaves")]
+    EmptyLeaves,
+}
 
 /// Simple MerkleTree.
 /// Reference:
@@ -11,26 +15,24 @@ use starknet_crypto::Felt;
 /// NOTE: Only supports the Pedersen hash for now.
 #[derive(Debug, Clone)]
 pub struct MerkleTree {
-    leaves: Vec<Felt>,
-    root_hash: Felt,
-    levels: Vec<Vec<Felt>>,
+    leaves: Vec<FieldElement>,
+    root_hash: FieldElement,
+    levels: Vec<Vec<FieldElement>>,
 }
 
 /// The merkle proof that a leaf belongs to a Merkle tree.
 #[derive(Debug, Clone, PartialEq, Default)]
-pub struct MerkleProof(pub Vec<Felt>);
+pub struct MerkleProof(pub Vec<FieldElement>);
 
 impl MerkleTree {
-    pub fn new(leaves: Vec<Felt>) -> Result<Self> {
+    pub fn new(leaves: Vec<FieldElement>) -> Result<Self, MerkleTreeError> {
         if leaves.is_empty() {
-            return Err(eyre!(
-                "Cannot build Merkle tree from an empty list of leaves."
-            ));
+            return Err(MerkleTreeError::EmptyLeaves);
         }
 
         let mut tree = MerkleTree {
             leaves,
-            root_hash: Felt::default(),
+            root_hash: FieldElement::default(),
             levels: Vec::new(),
         };
 
@@ -41,7 +43,7 @@ impl MerkleTree {
         Ok(tree)
     }
 
-    fn build(&self) -> (Felt, Vec<Vec<Felt>>) {
+    fn build(&self) -> (FieldElement, Vec<Vec<FieldElement>>) {
         if self.leaves.len() == 1 {
             return (self.leaves[0], vec![self.leaves.clone()]);
         }
@@ -60,7 +62,7 @@ impl MerkleTree {
                 let b = if chunk.len() > 1 {
                     chunk[1]
                 } else {
-                    Felt::ZERO
+                    FieldElement::ZERO
                 };
                 new_nodes.push(self.hash(&a, &b));
             }
@@ -74,7 +76,7 @@ impl MerkleTree {
         (curr_level_nodes[0], levels)
     }
 
-    fn hash(&self, a: &Felt, b: &Felt) -> Felt {
+    fn hash(&self, a: &FieldElement, b: &FieldElement) -> FieldElement {
         let (a_sorted, b_sorted) = match a.cmp(b) {
             Ordering::Less | Ordering::Equal => (a, b),
             Ordering::Greater => (b, a),
@@ -82,7 +84,7 @@ impl MerkleTree {
         pedersen_hash(a_sorted, b_sorted)
     }
 
-    pub fn get_proof(&self, leaf: &Felt) -> Option<MerkleProof> {
+    pub fn get_proof(&self, leaf: &FieldElement) -> Option<MerkleProof> {
         let mut path = Vec::new();
         let mut current_hash = *leaf;
 
@@ -93,7 +95,7 @@ impl MerkleTree {
             }
 
             let sibling_index = if index % 2 == 0 { index + 1 } else { index - 1 };
-            let sibling = level.get(sibling_index).unwrap_or(&Felt::ZERO);
+            let sibling = level.get(sibling_index).unwrap_or(&FieldElement::ZERO);
 
             path.push(*sibling);
             current_hash = self.hash(&current_hash, sibling);
@@ -101,7 +103,7 @@ impl MerkleTree {
         Some(MerkleProof(path))
     }
 
-    pub fn verify_proof(&self, leaf: &Felt, proof: &MerkleProof) -> bool {
+    pub fn verify_proof(&self, leaf: &FieldElement, proof: &MerkleProof) -> bool {
         let mut current_hash = *leaf;
         for &sibling in &proof.0 {
             current_hash = self.hash(&current_hash, &sibling);
@@ -118,7 +120,12 @@ mod tests {
 
     #[rstest]
     fn test_merkle_tree_new() {
-        let leaves = vec![Felt::from(1), Felt::from(2), Felt::from(3), Felt::from(4)];
+        let leaves = vec![
+            FieldElement::from(1_u32),
+            FieldElement::from(2_u32),
+            FieldElement::from(3_u32),
+            FieldElement::from(4_u32),
+        ];
 
         let merkle_tree = MerkleTree::new(leaves.clone()).unwrap();
 
@@ -126,25 +133,32 @@ mod tests {
         assert_eq!(merkle_tree.levels.len(), 3);
         assert_eq!(
             merkle_tree.root_hash,
-            Felt::from_hex_unchecked(
+            FieldElement::from_hex_be(
                 "0x38118a340bbba28e678413cd3b07a9436a5e60fd6a7cbda7db958a6d501e274"
             )
+            .unwrap()
         )
     }
 
     #[rstest]
     fn test_merkle_tree_proof() {
-        let leaves = vec![Felt::from(1), Felt::from(2), Felt::from(3), Felt::from(4)];
+        let leaves = vec![
+            FieldElement::from(1_u32),
+            FieldElement::from(2_u32),
+            FieldElement::from(3_u32),
+            FieldElement::from(4_u32),
+        ];
         let merkle_tree = MerkleTree::new(leaves.clone()).unwrap();
 
-        let leaf = Felt::from(1);
+        let leaf = FieldElement::from(1_u32);
         let proof = merkle_tree.get_proof(&leaf).unwrap();
 
         let expected_proof = MerkleProof(vec![
-            Felt::from_hex_unchecked("0x2"),
-            Felt::from_hex_unchecked(
+            FieldElement::from_hex_be("0x2").unwrap(),
+            FieldElement::from_hex_be(
                 "0x262697b88544f733e5c6907c3e1763131e9f14c51ee7951258abbfb29415fbf",
-            ),
+            )
+            .unwrap(),
         ]);
 
         assert_eq!(proof, expected_proof);
